Frontend: Moved dashboard and patient dashboard logic to features
