--- frontend/src/app/dashboard/patient/tests/[testId]/[assignedTestId]/components/onLeaveClick.ts
@@ -1,5 +1,0 @@
-import {redirect} from "next/navigation";
-
-const onLeaveClick = () => redirect("/dashboard/patient");
-
-export default onLeaveClick;
