+++ frontend/src/features/dashboard/patient/tests/[testId]/[assignedTestId]/components/LeaveTestButton.tsx
@@ -1,0 +1,15 @@
+import ArrowBackIcon from "@mui/icons-material/ArrowBack";
+import {Button} from "@mui/material";
+import {goToPatientPage} from "@/features/dashboard/patient/tests/[testId]/[assignedTestId]/utils";
+
+export default function LeaveTestButton() {
+    return (
+        <Button
+            onClick={goToPatientPage}
+            variant={"outlined"}
+            startIcon={<ArrowBackIcon />}
+        >
+            Покинути тест
+        </Button>
+    );
+}
