+++ frontend/src/features/dashboard/patient/tests/[testId]/[assignedTestId]/hooks/useTestContext.ts
@@ -1,0 +1,6 @@
+"use client";
+
+import {useContext} from "react";
+import TestContext from "@/features/dashboard/patient/tests/[testId]/[assignedTestId]/contexts/TestContext";
+
+export const useTestContext = () => useContext(TestContext);
