+++ frontend/src/features/dashboard/patient/tests/[testId]/[assignedTestId]/utils.ts
@@ -1,0 +1,3 @@
+import {redirect} from "next/navigation";
+
+export const goToPatientPage = () => redirect("/dashboard/patient");
